Updated 11/04
Included a deblurring where we use the true PSF as a benchmark for the optimal result we can expect. Also added a plot which show the convergence of the estimated blurring kernel parameter via r_update to the true one. Now also saves the figures in the script as images in the test folder, and extended the structure which contains the test results and parameters.